Remove unnecessary lexer_routines and rename lexer_trie to lexer
